Router class
Added the Graph creation in the constructor to check the isUpdated
attribute.
Included a new attribute called portNumber to define it in the
constructor and use it in the link.
Changed the link class but not all working.
Created the threads in the Program class.
Need to figure out a way to know when to start the submission of the
host's message.